+++ chempy/tests/ammonical_cupric_solution.py
@@ -2,8 +2,8 @@
 
 from __future__ import (absolute_import, division, print_function)
 
-from chempy.chemistry import Solute
-from chempy.equilibria import Equilibrium, EqSystemLog
+from chempy.chemistry import Solute, Equilibrium
+from chempy.equilibria import EqSystemLog
 import periodictable