+++ CHANGES.rst
@@ -2,6 +2,7 @@ v0.2.0
 ======
 - Support for solid phases in equilibria
 - Submodules for water properties moved to chempy.properties
+- Moved class ``Equilibrium`` from .equilibria to .chemistry
 
 v0.1
 ====